Working adding to the file, deleting from file doesn't work

+++ item.h
@@ -7,17 +7,17 @@ class Item
 {
 public:
     Item();
-    float getPrice() const {return m_price;}
+    int getPrice() const {return m_price;}
     QString getName() const {
         return m_name;
     }
     void setName(QString it) {m_name=it;}
-    void setPrice(float price) {m_price=price;}
+    void setPrice(int price) {m_price=price;}
     int getCount() const {return count;}
     void setCount(int c) {count = c;}
 private:
     QString m_name;
-    float m_price;
+    int m_price;
     int count=0;
 };
